Get SamplesToLocationsWorking

+++ src/api/songtrax.js
@@ -62,30 +62,47 @@ export async function getLocations(){
     return json;
 }
 
-
-export async function getSamplesToLocations(){
+export async function getSamplesToLocations(sampleID = -1){
     // Get
     const url = `${baseURL}sampletolocation/?api_key=${APIKEY}`;
     const response = await fetch(url);
     const json = await response.json();
 
-    return json;
+    if(sampleID === -1) return json;
+
+    json.filter(element => element.sample_id == sampleID);
+}
+ 
+async function getSampleToLocationIDFromIDs(sampleID, locationID){
+    // Get from API
+    const url = `${baseURL}sampletolocation/?api_key=${APIKEY}`;
+    const response = await fetch(url);
+    const json = await response.json(); 
+
+    // Filter to get only the record of matching IDs
+    const filteredResponse = json.filter(element => (element.sample_id === sampleID && element.location_id === locationID))
+    .map(element => element.id); 
+    return filteredResponse[0] ?? -1;
 }
 
-export async function removeSampleFromLocation(id){
-    const url = `${baseURL}sampletolocation/${id}/?api_key=${APIKEY}`;;
+export async function removeSampleFromLocation(sampleID, locationID){
+    const preexistingRelationshipID = await getSampleToLocationIDFromIDs(sampleID, locationID);
+    if(preexistingRelationshipID == -1) return null;
+
+    const url = `${baseURL}sampletolocation/${preexistingRelationshipID}/?api_key=${APIKEY}`;;
     await fetch(url, {
         method: 'DELETE',
         headers: {
             'Accept': 'application/json',
             'Content-Type': 'application/json'
-        },
-        body: {}
+        } 
     });
-}
-
+} 
 
 export async function addSampleToLocation(sampleID, locationID){
+    const preexistingRelationshipID = await getSampleToLocationIDFromIDs(sampleID, locationID);
+    if(preexistingRelationshipID != -1) return null;
+
     const url = `${baseURL}sampletolocation/?api_key=${APIKEY}`;
     const response = await fetch(url, {
         method: 'POST',
